requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
PyMuPDF
Pillow
openai